# cmd/swag/docs/swagger.yaml
basePath: /v1/2024
definitions:
  model.RegisterInput:
    properties:
      verify_key:
        type: string
      verify_purpose:
        type: string
      verify_type:
        type: integer
    type: object
  model.VerifyOTPInput:
    properties:
      verify_code:
        type: string
      verify_key:
        type: string
    type: object
  response.Response:
    properties:
      code:
        description: code infor for user
        type: integer
      data:
        description: dynamic data
      message:
        description: message infor for user
        type: string
    type: object
host: localhost:8090
info:
  contact:
    email: anhdung.phc@gmail.com
    name: Anh Dung Nguyen
    url: https://www.linkedin.com/in/anhdungnguyen-95262198/
  description: This is API Document for Ecommerce Backend Golang Project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/stephennguyen1803/ecom-golang
  title: API Ecommerce Backend Golang Project
  version: "1.0"
paths:
  /user/register:
    post:
      consumes:
      - application/json
      description: Register User Using Verify Key
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
      summary: Register User
      tags:
      - login
  /user/verify_account:
    post:
      consumes:
      - application/json
      description: Verfiy OTP login by user
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.VerifyOTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
      summary: Verify Account Using OTP
      tags:
      - login
schemes:
- http
swagger: "2.0"